requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiomysql>=0.2.0
pyyaml>=6.0.1
redis[hiredis]>=5.0.0
aioredis>=2.0.1
PyJWT>=2.8.0
strawberry-graphql[fastapi]>=0.200.0
watchdog>=3.0.0
httpx>=0.25.0
click>=8.1.0
prometheus-client>=0.19.0
python-dotenv>=1.0.0
structlog>=23.2.0

